Rename ./common to ./objects
