+++ src/model-interfaces/actors/actor.interface.ts
@@ -1,4 +1,4 @@
-import { APObject } from '../common/object.interface';
+import { APObject } from '../objects/object.interface';
 import {
     FollowersField,
     FollowingField,
@@ -12,7 +12,7 @@ import {
 } from '../../fields';
 
 /**
- * `Actor` types are [Object](../common/object.interface.ts)
+ * `Actor` types are [Object](../objects/object.interface.ts)
  * types that are capable of performing activities.
  * 
  * {@link https://www.w3.org/TR/activitystreams-vocabulary/#actor-types Docs}
