+++ src/model-interfaces/objects/event.interface.ts
