+++ src/model-interfaces/objects/link.interface.ts
