+++ src/model-interfaces/objects/object.interface.ts
