+++ src/model-interfaces/objects/place.interface.ts
