+++ src/model-interfaces/objects/relationship.interface.ts
