+++ src/model-interfaces/objects/tombstone.interface.ts
